+++ api/package.json
@@ -65,7 +65,7 @@
         "csv-stringify": "^6.3.0",
         "drizzle-kit": "^0.30.0",
         "drizzle-orm": "^0.38.0",
-        "drizzle-typebox": "^0.1.1",
+        "drizzle-typebox": "^0.2.0",
         "express": "^4.18.2",
         "express-minify": "^1.0.0",
         "fast-xml-parser": "^4.3.5",
